backend/requirements.txt
fastapi[standard]
numpy
pandas
torch
onnxruntime
Pillow
mlflow
onnx
